requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.4.0
auto-py-to-exe==2.17.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
loguru==0.6.0
outcome==1.1.0
pefile==2021.9.3
pycparser==2.21
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.1
pyOpenSSL==22.0.0
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.2
whichcraft==0.6.1
win32-setctime==1.1.0
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
